deploy/requirements.txt
snowflake-connector-python
snowflake-snowpark-python==1.42.0
dotenv